requirements.txt
absl-py==1.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
black==23.7.0
blinker==1.6.2
boto3==1.28.63
botocore==1.31.63
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.41.1
greenlet==3.0.0
h11==0.14.0
hydrofunctions==0.2.4
idna==3.4
importlib-metadata==6.8.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numpy==1.25.1
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
SQLAlchemy==2.0.21
stack-data==0.6.2
tk==0.1.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
webdriver-manager==4.0.0
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.16.2
